+++ changes.md
@@ -3,9 +3,19 @@
 <details>
 <summary>Added Type(s)</summary>
 
+- added type `CartMergeMode`
+- added type `MergeCartDraft`
 - added type `DiscountCombinationMode`
 - added type `DiscountsConfiguration`
 - added type `ProjectSetDiscountsConfigurationAction`
+</details>
+
+
+<details>
+<summary>Added Method(s)</summary>
+
+- added method `apiRoot.withProjectKey().carts().customerIdWithCustomerIdValueMerge().post()`
+- added method `apiRoot.withProjectKey().inStoreKeyWithStoreKeyValue().carts().customerIdWithCustomerIdValueMerge().post()`
 </details>
 
 
@@ -21,5 +31,13 @@
 
 - added property `discounts` to type `Project`
 - added property `skipConfigurationInputDraft` to type `RecurringOrderSetOrderSkipConfigurationAction`
+</details>
+
+
+<details>
+<summary>Added Resource(s)</summary>
+
+- added resource `/{projectKey}/carts/customer-id={customerId}/merge`
+- added resource `/{projectKey}/in-store/key={storeKey}/carts/customer-id={customerId}/merge`
 </details>